refactor: Replace scalar_conversion_to_int with blanket implementation
