+++ crates/proof-of-sql/src/base/scalar/mod.rs
@@ -89,173 +89,3 @@ pub trait Scalar:
         }
     }
 }
-
-macro_rules! scalar_conversion_to_int {
-    ($scalar:ty) => {
-        impl TryFrom<$scalar> for bool {
-            type Error = ScalarConversionError;
-            fn try_from(value: $scalar) -> Result<Self, Self::Error> {
-                let (sign, abs): (i128, [u64; 4]) = if value > <$scalar>::MAX_SIGNED {
-                    (-1, (-value).into())
-                } else {
-                    (1, value.into())
-                };
-                if abs[1] != 0 || abs[2] != 0 || abs[3] != 0 {
-                    return Err(ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i8",
-                        value
-                    )));
-                }
-                let val: i128 = sign * abs[0] as i128;
-                match val {
-                    0 => Ok(false),
-                    1 => Ok(true),
-                    _ => Err(ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in a bool",
-                        value
-                    ))),
-                }
-            }
-        }
-
-        impl TryFrom<$scalar> for i8 {
-            type Error = ScalarConversionError;
-            fn try_from(value: $scalar) -> Result<Self, Self::Error> {
-                let (sign, abs): (i128, [u64; 4]) = if value > <$scalar>::MAX_SIGNED {
-                    (-1, (-value).into())
-                } else {
-                    (1, value.into())
-                };
-                if abs[1] != 0 || abs[2] != 0 || abs[3] != 0 {
-                    return Err(ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i8",
-                        value
-                    )));
-                }
-                let val: i128 = sign * abs[0] as i128;
-                val.try_into().map_err(|_| {
-                    ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i8",
-                        value
-                    ))
-                })
-            }
-        }
-
-        impl TryFrom<$scalar> for i16 {
-            type Error = ScalarConversionError;
-            fn try_from(value: $scalar) -> Result<Self, Self::Error> {
-                let (sign, abs): (i128, [u64; 4]) = if value > <$scalar>::MAX_SIGNED {
-                    (-1, (-value).into())
-                } else {
-                    (1, value.into())
-                };
-                if abs[1] != 0 || abs[2] != 0 || abs[3] != 0 {
-                    return Err(ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i16",
-                        value
-                    )));
-                }
-                let val: i128 = sign * abs[0] as i128;
-                val.try_into().map_err(|_| {
-                    ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i16",
-                        value
-                    ))
-                })
-            }
-        }
-
-        impl TryFrom<$scalar> for i32 {
-            type Error = ScalarConversionError;
-            fn try_from(value: $scalar) -> Result<Self, Self::Error> {
-                let (sign, abs): (i128, [u64; 4]) = if value > <$scalar>::MAX_SIGNED {
-                    (-1, (-value).into())
-                } else {
-                    (1, value.into())
-                };
-                if abs[1] != 0 || abs[2] != 0 || abs[3] != 0 {
-                    return Err(ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i32",
-                        value
-                    )));
-                }
-                let val: i128 = sign * abs[0] as i128;
-                val.try_into().map_err(|_| {
-                    ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i32",
-                        value
-                    ))
-                })
-            }
-        }
-
-        impl TryFrom<$scalar> for i64 {
-            type Error = ScalarConversionError;
-            fn try_from(value: $scalar) -> Result<Self, Self::Error> {
-                let (sign, abs): (i128, [u64; 4]) = if value > <$scalar>::MAX_SIGNED {
-                    (-1, (-value).into())
-                } else {
-                    (1, value.into())
-                };
-                if abs[1] != 0 || abs[2] != 0 || abs[3] != 0 {
-                    return Err(ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i64",
-                        value
-                    )));
-                }
-                let val: i128 = sign * abs[0] as i128;
-                val.try_into().map_err(|_| {
-                    ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i64",
-                        value
-                    ))
-                })
-            }
-        }
-
-        impl TryFrom<$scalar> for i128 {
-            type Error = ScalarConversionError;
-            fn try_from(value: $scalar) -> Result<Self, Self::Error> {
-                let (sign, abs): (i128, [u64; 4]) = if value > <$scalar>::MAX_SIGNED {
-                    (-1, (-value).into())
-                } else {
-                    (1, value.into())
-                };
-                if abs[2] != 0 || abs[3] != 0 {
-                    return Err(ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i128",
-                        value
-                    )));
-                }
-                let val: u128 = (abs[1] as u128) << 64 | (abs[0] as u128);
-                match (sign, val) {
-                    (1, v) if v <= i128::MAX as u128 => Ok(v as i128),
-                    (-1, v) if v <= i128::MAX as u128 => Ok(-(v as i128)),
-                    (-1, v) if v == i128::MAX as u128 + 1 => Ok(i128::MIN),
-                    _ => Err(ScalarConversionError::Overflow(format!(
-                        "{} is too large to fit in an i128",
-                        value
-                    ))),
-                }
-            }
-        }
-
-        impl From<$scalar> for BigInt {
-            fn from(value: $scalar) -> Self {
-                // Since we wrap around in finite fields anything greater than the max signed value is negative
-                let is_negative = value > <$scalar>::MAX_SIGNED;
-                let sign = if is_negative {
-                    num_bigint::Sign::Minus
-                } else {
-                    num_bigint::Sign::Plus
-                };
-                let value_abs: [u64; 4] = (if is_negative { -value } else { value }).into();
-                let bits: &[u8] = bytemuck::cast_slice(&value_abs);
-                BigInt::from_bytes_le(sign, &bits)
-            }
-        }
-    };
-}
-
-pub(crate) use scalar_conversion_to_int;
